add new test class to cover the mandatory letter for adminAppealWithdrawal
