+++ src/IntegrationTests/java/uk/gov/hmcts/reform/sscs/service/NotificationServiceForSubscriptionUpdatedTest.java
@@ -5,37 +5,16 @@ import static org.mockito.Mockito.eq;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoMoreInteractions;
-import static uk.gov.hmcts.reform.sscs.ccd.domain.EventType.APPEAL_RECEIVED;
 import static uk.gov.hmcts.reform.sscs.domain.notify.NotificationEventType.ADMIN_APPEAL_WITHDRAWN;
 import static uk.gov.hmcts.reform.sscs.domain.notify.NotificationEventType.SUBSCRIPTION_UPDATED_NOTIFICATION;
 
-import java.util.ArrayList;
-import java.util.List;
 import junitparams.Parameters;
 import org.junit.Ignore;
 import org.junit.Test;
 import org.springframework.beans.factory.annotation.Value;
-import uk.gov.hmcts.reform.sscs.ccd.domain.Address;
-import uk.gov.hmcts.reform.sscs.ccd.domain.Appeal;
-import uk.gov.hmcts.reform.sscs.ccd.domain.AppealReasons;
-import uk.gov.hmcts.reform.sscs.ccd.domain.Appellant;
-import uk.gov.hmcts.reform.sscs.ccd.domain.Appointee;
-import uk.gov.hmcts.reform.sscs.ccd.domain.Benefit;
-import uk.gov.hmcts.reform.sscs.ccd.domain.BenefitType;
-import uk.gov.hmcts.reform.sscs.ccd.domain.Contact;
-import uk.gov.hmcts.reform.sscs.ccd.domain.Event;
-import uk.gov.hmcts.reform.sscs.ccd.domain.EventDetails;
-import uk.gov.hmcts.reform.sscs.ccd.domain.HearingOptions;
-import uk.gov.hmcts.reform.sscs.ccd.domain.Identity;
-import uk.gov.hmcts.reform.sscs.ccd.domain.MrnDetails;
-import uk.gov.hmcts.reform.sscs.ccd.domain.Name;
-import uk.gov.hmcts.reform.sscs.ccd.domain.Representative;
 import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseData;
 import uk.gov.hmcts.reform.sscs.ccd.domain.Subscription;
-import uk.gov.hmcts.reform.sscs.ccd.domain.Subscriptions;
-import uk.gov.hmcts.reform.sscs.config.AppealHearingType;
 import uk.gov.hmcts.reform.sscs.domain.SscsCaseDataWrapper;
-import uk.gov.hmcts.reform.sscs.domain.notify.NotificationEventType;
 import uk.gov.hmcts.reform.sscs.factory.CcdNotificationWrapper;
 import uk.gov.service.notify.NotificationClientException;
 
@@ -79,21 +58,6 @@ public class NotificationServiceForSubscriptionUpdatedTest extends NotificationS
 
     @Value("${notification.subscriptionOld.smsId}")
     private String subscriptionOldSmsId;
-
-    @Test
-    @Ignore
-    public void adminAppealWithdrawalWhenNoSubscription_shouldSendMandatoryLetter() throws Exception {
-        SscsCaseData newSscsCaseData = getSscsCaseData(null);
-        SscsCaseDataWrapper wrapper = getSscsCaseDataWrapper(newSscsCaseData, null, ADMIN_APPEAL_WITHDRAWN);
-
-        getNotificationService().manageNotificationAndSubscription(new CcdNotificationWrapper(wrapper));
-
-        verify(getNotificationSender(), times(0)).sendEmail(any(), any(), any(), any(), any(), any());
-        verify(getNotificationSender(), times(0))
-            .sendSms(any(), any(), any(), any(), any(), any(), any());
-
-
-    }
 
     @Test
     @Parameters({"appellant", "representative", "appointee"})
@@ -377,14 +341,6 @@ public class NotificationServiceForSubscriptionUpdatedTest extends NotificationS
         verifyNoMoreInteractions(getNotificationSender());
     }
 
-    private SscsCaseDataWrapper getSscsCaseDataWrapper(SscsCaseData newSscsCaseData, SscsCaseData oldSscsCaseData,
-                                                       NotificationEventType subscriptionUpdatedNotification) {
-        return SscsCaseDataWrapper.builder()
-            .newSscsCaseData(newSscsCaseData)
-            .oldSscsCaseData(oldSscsCaseData)
-            .notificationEventType(subscriptionUpdatedNotification).build();
-    }
-
     private String getAppealReceivedEmailId(String who) {
         if (who.equals("appellant")) {
             return appealReceivedAppellantEmailId;
@@ -403,71 +359,5 @@ public class NotificationServiceForSubscriptionUpdatedTest extends NotificationS
         return appealReceivedAppointeeSmsId;
     }
 
-    private SscsCaseData getSscsCaseData(Subscription subscription, String who) {
-        if (who.equals("appellant")) {
-            return getSscsCaseData(subscription);
-        } else if (who.equals("representative")) {
-            return getSscsCaseDataForRep(subscription);
-        }
-        return getSscsCaseDataForAppointee(subscription);
-    }
 
-    private SscsCaseData getSscsCaseData(Subscription subscription) {
-        List<Event> events = new ArrayList<>();
-        events.add(Event.builder().value(EventDetails.builder().date(NotificationServiceBase.DATE).type(APPEAL_RECEIVED.getCcdType()).build()).build());
-
-        return SscsCaseData.builder().ccdCaseId(NotificationServiceBase.CASE_ID).events(events)
-            .appeal(Appeal.builder()
-                .mrnDetails(MrnDetails.builder().mrnDate(NotificationServiceBase.DATE).dwpIssuingOffice("office").build())
-                .appealReasons(AppealReasons.builder().build())
-                .rep(Representative.builder()
-                    .hasRepresentative(NotificationServiceBase.YES)
-                    .name(Name.builder().firstName("Rep").lastName("lastName").build())
-                    .contact(Contact.builder().email(NotificationServiceBase.EMAIL_TEST_2).phone(NotificationServiceBase.MOBILE_NUMBER_2).build())
-                    .address(Address.builder().line1("Rep Line 1").town("Rep Town").county("Rep County").postcode("RE9 7SE").build())
-                    .build())
-                .appellant(Appellant.builder()
-                    .name(Name.builder().firstName("firstName").lastName("lastName").build())
-                    .address(Address.builder().line1("122 Breach Street").line2("The Village").town("My town").county("Cardiff").postcode("CF11 2HB").build())
-                    .contact(Contact.builder().email(NotificationServiceBase.EMAIL_TEST_1).phone(NotificationServiceBase.MOBILE_NUMBER_1).build())
-                    .identity(Identity.builder().nino("NP 27 28 67 B").dob("12 March 1971").build()).build())
-                .hearingType(AppealHearingType.ORAL.name())
-                .benefitType(BenefitType.builder().code(Benefit.PIP.name()).build())
-                .hearingOptions(HearingOptions.builder()
-                    .wantsToAttend(NotificationServiceBase.YES)
-                    .build())
-                .build())
-            .subscriptions(Subscriptions.builder()
-                .appellantSubscription(subscription)
-                .representativeSubscription(getSubscription().toBuilder().tya("REP_TYA").build())
-                .build())
-            .caseReference(NotificationServiceBase.CASE_REFERENCE).build();
-    }
-
-    private SscsCaseData getSscsCaseDataForRep(Subscription subscription) {
-        Subscription appellantSubscription = getSubscription().toBuilder().tya("APPELLANT_TYA").build();
-        SscsCaseData sscsCaseData = getSscsCaseData(appellantSubscription);
-        return sscsCaseData.toBuilder()
-            .subscriptions(sscsCaseData.getSubscriptions().toBuilder().representativeSubscription(subscription).build())
-            .build();
-    }
-
-    private SscsCaseData getSscsCaseDataForAppointee(Subscription subscription) {
-        SscsCaseData sscsCaseData = getSscsCaseData(subscription);
-        return sscsCaseData.toBuilder()
-            .appeal(sscsCaseData.getAppeal().toBuilder().appellant(sscsCaseData.getAppeal().getAppellant()
-
-                .toBuilder()
-                .appointee(Appointee.builder()
-                    .name(Name.builder().firstName("Appoin").lastName("Tee").build())
-                    .address(sscsCaseData.getAppeal().getAppellant().getAddress())
-                    .contact(sscsCaseData.getAppeal().getAppellant().getContact())
-                    .identity(sscsCaseData.getAppeal().getAppellant().getIdentity())
-                    .build())
-                .build()).build())
-            .subscriptions(sscsCaseData.getSubscriptions().toBuilder()
-                .appellantSubscription(null)
-                .appointeeSubscription(subscription).build())
-            .build();
-    }
 }
